+++ week2/demo/demo/.gitignore
@@ -38,3 +38,4 @@ out/
 
 g### application.yml ###
 application.yml
+application-dev.yml
