--- week2/demo/demo/src/main/java/com/sopt/demo/dto/request/MemberPatchRequest.java
@@ -1,7 +1,0 @@
-package com.sopt.demo.dto.request;
-
-import com.sopt.demo.domain.Part;
-import com.sopt.demo.domain.SOPT;
-
-public record MemberPatchRequest(String nickname, int age, SOPT sopt) {
-}
