--- week2/demo/demo/src/main/java/com/sopt/demo/dto/request/MemberPutRequest.java
@@ -1,7 +1,0 @@
-package com.sopt.demo.dto.request;
-
-import com.sopt.demo.domain.SOPT;
-
-
-public record MemberPutRequest(String nickname, int age, SOPT sopt) {
-}
