+++ week2/demo/demo/src/main/java/com/sopt/demo/dto/request/member/MemberPatchRequest.java
@@ -1,0 +1,6 @@
+package com.sopt.demo.dto.request.member;
+
+import com.sopt.demo.domain.member.SOPT;
+
+public record MemberPatchRequest(String nickname, int age, SOPT sopt) {
+}
