+++ week2/demo/demo/src/main/java/com/sopt/demo/dto/request/member/MemberPutRequest.java
@@ -1,0 +1,7 @@
+package com.sopt.demo.dto.request.member;
+
+import com.sopt.demo.domain.member.SOPT;
+
+
+public record MemberPutRequest(String nickname, int age, SOPT sopt) {
+}
