+++ week2/demo/demo/src/main/java/com/sopt/demo/dto/response/category/CategoryResponse.java
@@ -1,6 +1,6 @@
 package com.sopt.demo.dto.response.category;
 
-import com.sopt.demo.domain.Category;
+import com.sopt.demo.domain.category.Category;
 import lombok.Data;
 
 @Data
