+++ week2/demo/demo/src/main/java/com/sopt/demo/dto/response/post/PostGetResponse.java
@@ -1,7 +1,7 @@
 package com.sopt.demo.dto.response.post;
 
-import com.sopt.demo.domain.Category;
-import com.sopt.demo.domain.Post;
+import com.sopt.demo.domain.category.Category;
+import com.sopt.demo.domain.post.Post;
 
 public record PostGetResponse(Long id, String title, String content, String category) {
     public static PostGetResponse of(Post post, Category category) {
