+++ week2/demo/demo/src/main/java/com/sopt/demo/repository/CategoryJpaRepository.java
@@ -1,6 +1,6 @@
 package com.sopt.demo.repository;
 
-import com.sopt.demo.domain.Category;
+import com.sopt.demo.domain.category.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CategoryJpaRepository extends JpaRepository<Category, Short> {
