+++ week2/demo/demo/src/main/resources/application-dev.yml
@@ -1,0 +1,15 @@
+spring:
+  datasource:
+    driver-class-name: com.mysql.cj.jdbc.Driver
+    url: jdbc:mysql://localhost:3306/dosopt?serverTimezone=UTC&characterEncoding=UTF-8
+    username: root
+    password: soptseminar!
+
+  config:
+    activate:
+      on-profile: dev
+
+  jpa:
+    database-platform: org.hibernate.dialect.MySQL8Dialect
+    hibernate:
+      ddl-auto: create
